2025-2/IA/IA Gato e Cachorro/requirements.txt
# Dependências para o projeto IA Classificadora de Gatos e Cachorros

# Core ML
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# Processamento de imagens
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.19.0

# Interface web
streamlit>=1.0.0

# Utilitários
joblib>=1.0.0
python-dotenv>=0.19.0

# Desenvolvimento (opcional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0
